Refactor project removed ide.
