+++ com.arcbees.ide.plugin.eclipse/.classpath
@@ -20,11 +20,10 @@
 	<classpathentry exported="true" kind="lib" path="lib/httpmime-4.2.1.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/rest-assured-1.8.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/tagsoup-1.2.1.jar"/>
-	<classpathentry kind="lib" path="lib/commons-configuration-1.9.jar"/>
-	<classpathentry kind="lib" path="lib/swing2swt.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/commons-configuration-1.9.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/swing2swt.jar"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="/Users/branflake2267/git/gwtp-eclipse-plugin/com.arcbees.ide.plugin.eclipse/lib/swing2swt.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
