+++ com.arcbees.ide.plugin.eclipse/plugin.xml
@@ -6,39 +6,39 @@
          point="org.eclipse.ui.commands">
       <category
             name="Sample Category"
-            id="com.arcbees.ide.plugin.eclipse.commands.category">
+            id="com.arcbees.plugin.eclipse.commands.category">
       </category>
       <command
             name="Create Project"
-            categoryId="com.arcbees.ide.plugin.eclipse.commands.category"
-            id="com.arcbees.ide.plugin.eclipse.commands.createProject">
+            categoryId="com.arcbees.plugin.eclipse.commands.category"
+            id="com.arcbees.plugin.eclipse.commands.createProject">
       </command>
       <command
-            defaultHandler="com.arcbees.ide.plugin.eclipse.handlers.CreateProjectHandler"
+            defaultHandler="com.arcbees.plugin.eclipse.handlers.CreateProjectHandler"
             id="referenceToMenuId"
             name="referenceToMenuId">
       </command>
       <command
-            categoryId="com.arcbees.ide.plugin.eclipse.commands.category"
-            id="com.arcbees.ide.plugin.eclipse.commands.createPresenter"
+            categoryId="com.arcbees.plugin.eclipse.commands.category"
+            id="com.arcbees.plugin.eclipse.commands.createPresenter"
             name="Create Presenter">
       </command>
    </extension>
    <extension
          point="org.eclipse.ui.handlers">
       <handler
-            commandId="com.arcbees.ide.plugin.eclipse.commands.createProject"
-            class="com.arcbees.ide.plugin.eclipse.handlers.CreateProjectHandler">
+            commandId="com.arcbees.plugin.eclipse.commands.createProject"
+            class="com.arcbees.plugin.eclipse.handlers.CreateProjectHandler">
       </handler>
       <handler
-            class="com.arcbees.ide.plugin.eclipse.handlers.CreatePresenterHandler"
-            commandId="com.arcbees.ide.plugin.eclipse.commands.createPresenter">
+            class="com.arcbees.plugin.eclipse.handlers.CreatePresenterHandler"
+            commandId="com.arcbees.plugin.eclipse.commands.createPresenter">
       </handler>
    </extension>
    <extension
          point="org.eclipse.ui.bindings">
       <key
-            commandId="com.arcbees.ide.plugin.eclipse.commands.createProject"
+            commandId="com.arcbees.plugin.eclipse.commands.createProject"
             contextId="org.eclipse.ui.contexts.window"
             sequence="M1+6"
             schemeId="org.eclipse.ui.defaultAcceleratorConfiguration">
@@ -49,7 +49,7 @@
       <menuContribution
             locationURI="toolbar:org.eclipse.ui.main.toolbar">
          <toolbar
-               id="com.arcbees.ide.plugin.eclipse.toolbars.gwtp">
+               id="com.arcbees.plugin.eclipse.toolbars.gwtp">
             <command
                   commandId="referenceToMenuId"
                   icon="icons/addgwtp.png"
@@ -62,12 +62,12 @@
             allPopups="false"
             locationURI="menu:referenceToMenuId">
          <command
-               commandId="com.arcbees.ide.plugin.eclipse.commands.createProject"
+               commandId="com.arcbees.plugin.eclipse.commands.createProject"
                label="Create GWTP Project"
                style="push">
          </command>
          <command
-               commandId="com.arcbees.ide.plugin.eclipse.commands.createPresenter"
+               commandId="com.arcbees.plugin.eclipse.commands.createPresenter"
                label="Create Presenter"
                style="push">
          </command>
