--- com.arcbees.ide.plugin.eclipse/src/com/arcbees/ide/plugin/eclipse/handlers/CreatePresenterHandler.java
@@ -1,46 +1,0 @@
-/**
- * Copyright 2013 ArcBees Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package com.arcbees.ide.plugin.eclipse.handlers;
-
-import org.eclipse.core.commands.AbstractHandler;
-import org.eclipse.core.commands.ExecutionEvent;
-import org.eclipse.core.commands.ExecutionException;
-import org.eclipse.jface.wizard.Wizard;
-import org.eclipse.jface.wizard.WizardDialog;
-import org.eclipse.ui.IWorkbenchWindow;
-import org.eclipse.ui.handlers.HandlerUtil;
-
-import com.arcbees.ide.plugin.eclipse.wizard.createpresenter.CreatePresenterWizard;
-
-public class CreatePresenterHandler extends AbstractHandler {
-    public CreatePresenterHandler() {
-    }
-
-    public Object execute(ExecutionEvent event) throws ExecutionException {
-        IWorkbenchWindow window = HandlerUtil.getActiveWorkbenchWindowChecked(event);
-        open(window);
-        return null;
-    }
-
-    private void open(IWorkbenchWindow window) {
-        Wizard wizard = new CreatePresenterWizard();
-       
-        WizardDialog dialog = new WizardDialog(window.getShell(), wizard);
-        dialog.create();
-        dialog.open();
-    }    
-}
